server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
